Conflicts:
	dotCMS/WEB-INF/jsp/license/license.jsp
	dotCMS/WEB-INF/lib/ee-f0d9d6fd1a61c5.jar
	src/com/dotmarketing/util/TaskLocatorUtil.java

Former-commit-id: 295ac0fd55f092880cf65dca22840669d46b69f1

+++ src/com/dotmarketing/servlets/InitServlet.java
@@ -178,9 +178,13 @@ public class InitServlet extends HttpServlet {
         new java.io.File(ConfigUtils.getDynamicVelocityPath() + File.separator + "live").mkdirs();
         new java.io.File(ConfigUtils.getDynamicVelocityPath() + File.separator + "working").mkdirs();
 
-        // deletes all menues that have been generated
-        RefreshMenus.deleteMenus();
-        CacheLocator.getNavToolCache().clearCache();
+
+        
+        if(Config.getBooleanProperty("CACHE_DISK_SHOULD_DELETE_NAVTOOL", false)){
+            // deletes all menues that have been generated
+            RefreshMenus.deleteMenus();
+        	CacheLocator.getCacheAdministrator().flushGroupLocalOnly("navCache");
+        }
 
 
         // maps all virtual links in memory
